% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasing_algorithm.R
\name{BC}
\alias{BC}
\title{Compute the bias correction for the unbiased estimators}
\usage{
BC(c_chains, h = function(x) x, k = 0, K = 1)
}
\description{
Compute the bias correction, for each of the element
in the list 'c_chains'. The integral of interest is that of the function h,
which can be multivariate. The estimator uses the variance reduction technique
whereby the estimator is the MCMC average between times k and K, with probability
going to one as k increases. In the case that K=Inf, the returned value is the 
almost sure limit of the bias correction
}
