% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoupledPIMH-package.R
\docType{package}
\name{CoupledPIMH-package}
\alias{CoupledPIMH-package}
\alias{CoupledPIMH}
\title{CoupledPIMH}
\description{
...
}
\details{
...
}
\author{
Lawrence T. Middleton <lawrence.middleton.work@gmail.com>
}
\keyword{package}
