% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_constrained_diffusion.R
\name{get_constrained_diffusion}
\alias{get_constrained_diffusion}
\title{Constrained sine diffusion}
\usage{
get_constrained_diffusion(time_T, T_final)
}
\value{
A list
}
\description{
This function returns a list with objects such as
* rinit, rinit_rand to sample from the initial distribution
* rtransition, rtransition_rand to sample from the transition
* dtransition to evaluate the transition density
* dmeasurement to evaluate the measurement density
* dimension, which represents the dimension of the latent process
}
