% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_kernels.R
\name{get_cpf_kernels}
\alias{get_cpf_kernels}
\title{CCPF kernels}
\usage{
get_cpf_kernels(with_as)
}
\arguments{
\item{with_as}{whether ancestor sampling should be used (TRUE/FALSE)}
}
\value{
Two kernels - the single kernel and the coupled transition kernel
}
\description{
Gets the kernels for CCPF transitions
}
