% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_kernels.R
\name{get_csmc_refresh_kernels}
\alias{get_csmc_refresh_kernels}
\title{CSMC refresh kernels}
\usage{
get_csmc_refresh_kernels(with_as)
}
\arguments{
\item{with_as}{whether ancestor sampling should be used (TRUE/FALSE)}
}
\value{
Two kernels - the single kernel and the coupled transition kernel
}
\description{
Gets the kernels for CSMC refresh
}
