% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_levy_sv.R
\name{get_levy_sv}
\alias{get_levy_sv}
\title{Levy-driven stochastic volatility model as in Barndorff-Nielsen and Shephard (2001)}
\usage{
get_levy_sv()
}
\value{
A list
}
\description{
This function returns a list with objects such as
* rinit, rinit_rand to sample from the initial distribution
* rtransition, rtransition_rand to sample from the transition
* dtransition to evaluate the transition density
* dmeasurement to evaluate the measurement density
* dimension, which represents the dimension of the latent process
}
