% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_kernels.R
\name{get_pimh_kernels}
\alias{get_pimh_kernels}
\title{PIMH kernels}
\usage{
get_pimh_kernels(with_as, all_particles = F)
}
\arguments{
\item{with_as}{whether ancestor sampling should be used (TRUE/FALSE)}

\item{all_particles}{whether to return the expected path using all the particles}
}
\value{
Two kernels - the single kernel and the coupled transition kernel
}
\description{
Gets the kernels for PIMH transitions
}
