% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_kernels.R
\name{get_pimh_smc_sampler_kernels}
\alias{get_pimh_smc_sampler_kernels}
\title{PIMH kernels for SMC sampler}
\usage{
get_pimh_smc_sampler_kernels(h_func = NULL)
}
\arguments{
\item{h_func}{If specified, this will be the Rao-Blackwellised test function}
}
\value{
Two kernels - the single kernel and the coupled transition kernel
}
\description{
Gets the kernels for PIMH transitions
}
