% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPF.R
\name{smc_sampler}
\alias{smc_sampler}
\title{SMC sampler}
\usage{
smc_sampler(nparticles, model, theta, systematic = FALSE,
  h_func = NULL)
}
\arguments{
\item{nparticles}{number of particles}

\item{model}{initial and transition densities for HMM}

\item{theta}{model parameters}
}
\description{
runs an SMC sampler with specified model, model parameters and number of particles
}
